Work done in first half of today
A number of to-dos done in Basecamp. Here's copypasta from Vitamin-R pomodoro completed description:

"Have the GmailReg class use the data from database when applicable

Figuring out how to refactor and minimize all the code repeating between Gmail, Yahoo, Aol.  Try not to spend more than a few pomodoros on this and def not the only work you do today.

For the refactoring and DRY, was able to grok code more while doing a bit of ultimately un-necessary DRYing. But overall ended up being a good idea.
- Made ‘check_domain’ and ‘finishing_up_new_email’ methods and placed in the Email superclass.
- Created a ‘provider’ variable that just says what sort of provider/domain the current object is of

A few other small to-dos done. Like:
- getting agem.rb required when not using Rails
- organizing gems/Gemfiles in aProject and Bing Rewards.
- And a bug fix like of the requiring of the zip code file in Generation not being good"